+++ src/lib.rs
@@ -52,7 +52,6 @@
 #![deny(broken_intra_doc_links)]
 #![deny(missing_debug_implementations)]
 #![deny(missing_docs)]
-#![deny(unsafe_code)]
 
 #[macro_use]
 mod macros;